requirements.txt
django-ckeditor
pillow
django-imagekit
django-uuslug
